updated naming conventions
