+++ scripts/main.js
@@ -1,4 +1,4 @@
 // Authors : Chase, John, Max, and Paul
 // Purpose : ¯\_(ツ)_/¯
 const Database = require("./Database");
-const registration_login = require("./usersCreator");
+const registration_login = require("./usersRegister");
